Modulariza e estrutura melhor o código do bot
